+++ src/hooks/useAnchoredPosition.ts
@@ -3,7 +3,7 @@
 import type { CSSProperties, RefObject } from 'react'
 import { useCallback, useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react'
 import { MathUtil } from '@/src/utils/math'
-import { useWindowResizeObserver } from './useResizeCallbackWrapper'
+import { useWindowResizeObserver } from './useWindowResizeObserver'
 import { useResizeObserver } from './useResizeObserver'
 import { useScrollObserver } from './useScrollObserver'